+++ CHANGELOG.md
@@ -12,4 +12,5 @@
 
   * Fixed pipeline path is none error ([#16](https://github.com/KIT-IAI/pywatts-pipeline/issues/16))
   * Fixed summary step crashing ([#25](https://github.com/KIT-IAI/pywatts-pipeline/issues/25))
+  * Pipeline result and step buffer are reset after training ([#35](https://github.com/KIT-IAI/pywatts-pipeline/issues/35))